out check health to separate class
